requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-jazzmin==3.0.0
django-minio-storage==0.5.7
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
idna==3.7
inflection==0.5.1
minio==7.2.7
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-dotenv-vault==0.6.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
